[eyes-cypress] add type definition for eyes plugin (#1171)
- add support for `ESM` and `TS`
- add support to `xvfb` using the `github action` `test`
- change the `plugin` to `typescript`
- tests to `cypress` `9`, `10`, `11`, `12`
- add support to `moduleResolution:node` using `ttsc`

+++ js/packages/eyes-cypress/test/unit/plugin/config.test.js
@@ -1,7 +1,7 @@
 'use strict'
 const {describe, it} = require('mocha')
 const {expect} = require('chai')
-const makeConfig = require('../../../src/plugin/config')
+const makeConfig = require('../../../dist/plugin/config').default
 const fs = require('fs')
 const path = require('path')
 
@@ -41,33 +41,33 @@ describe('config', () => {
     delete process.env.APPLITOOLS_BATCH_ID
   })
 
-  it('should convert accessibilityValidation to acceessibilityValidation', () => {
-    const filePath = path.join(__dirname, '../../../applitools.config.js')
-    fs.writeFileSync(filePath, "module.exports = {accessibilityValidation: 'AA'};")
-    const {config} = makeConfig()
-    try {
-      expect(config.accessibilitySettings).to.equal('AA')
-    } finally {
+  const filePath = path.join(__dirname, '../../../applitools.config.js')
+
+  afterEach(() => {
+    if (fs.existsSync(filePath)) {
       fs.unlinkSync(filePath)
     }
   })
 
+  it('should convert accessibilityValidation to acceessibilityValidation', () => {
+    fs.writeFileSync(filePath, "module.exports = {accessibilityValidation: 'AA'};")
+    const {config} = makeConfig()
+    expect(config.accessibilitySettings).to.equal('AA')
+  })
+
   it('should create random batch id when batch id is not defined in config file', () => {
-    const filePath = path.join(__dirname, '../../../applitools.config.js')
     fs.writeFileSync(filePath, 'module.exports = {};')
     const {config} = makeConfig()
     expect(config.batch.id).not.undefined
   })
 
   it('should not overwrite batch id from config file when passed in an object', () => {
-    const filePath = path.join(__dirname, '../../../applitools.config.js')
     fs.writeFileSync(filePath, "module.exports = {batch: {id: '1234'}};")
     const {config} = makeConfig()
     expect(config.batch.id).to.equal('1234')
   })
 
   it('should not overwrite batch id from config file when passed in as a property', () => {
-    const filePath = path.join(__dirname, '../../../applitools.config.js')
     fs.writeFileSync(filePath, "module.exports = {batchId: '1234'};")
     const {config} = makeConfig()
     expect(config.batch).to.be.undefined
@@ -75,7 +75,6 @@ describe('config', () => {
   })
 
   it('should not overwrite batch name from config file when passed in as an object', () => {
-    const filePath = path.join(__dirname, '../../../applitools.config.js')
     fs.writeFileSync(filePath, "module.exports = {batch: {name: '1234'}};")
     const {config} = makeConfig()
     expect(config.batch.name).to.be.equal('1234')
